a lot of visualization plots and stuff, check it out. Includes model mentioned in google doc
This model works and so do the plotting libraries. You can change the required file paths and change some model parameters and test it out. I have written this for the entire dataset. You guys can change the output of the fully connected layer and test it out on individual data folders